-Added "ActivityComponent"
