+++ app/src/main/java/com/mohamed/medhat/graduation_project/dagger/DaggerApplication.kt
@@ -1,7 +1,12 @@
 package com.mohamed.medhat.graduation_project.dagger
 
 import android.app.Application
+import com.mohamed.medhat.graduation_project.dagger.components.AppComponent
+import com.mohamed.medhat.graduation_project.dagger.components.DaggerAppComponent
 
+/**
+ * A dagger application.
+ */
 class DaggerApplication : Application() {
     // Instance of the AppComponent that will be used by all the Activities in the project
     val appComponent: AppComponent by lazy {
