+++ app/src/main/java/com/mohamed/medhat/graduation_project/dagger/components/ActivityComponent.kt
@@ -1,0 +1,22 @@
+package com.mohamed.medhat.graduation_project.dagger.components
+
+import com.mohamed.medhat.graduation_project.dagger.scopes.ActivityScope
+import com.mohamed.medhat.graduation_project.ui.login_activity.LoginActivity
+import com.mohamed.medhat.graduation_project.ui.splash_activity.SplashActivity
+import dagger.Subcomponent
+
+/**
+ * A dagger subcomponent for the activity lifecycle.
+ */
+@ActivityScope
+@Subcomponent
+interface ActivityComponent {
+
+    @Subcomponent.Factory
+    interface Factory {
+        fun create(): ActivityComponent
+    }
+
+    fun inject(loginActivity: LoginActivity)
+    fun inject(splashActivity: SplashActivity)
+}
