+++ app/src/main/java/com/mohamed/medhat/graduation_project/dagger/scopes/ActivityScope.kt
@@ -1,0 +1,8 @@
+package com.mohamed.medhat.graduation_project.dagger.scopes
+
+import javax.inject.Scope
+
+@Scope
+@MustBeDocumented
+@Retention(value = AnnotationRetention.RUNTIME)
+annotation class ActivityScope()
