+++ app/src/main/java/com/mohamed/medhat/graduation_project/ui/base/BaseActivity.kt
@@ -2,17 +2,18 @@ package com.mohamed.medhat.graduation_project.ui.base
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.mohamed.medhat.graduation_project.dagger.AppComponent
 import com.mohamed.medhat.graduation_project.dagger.DaggerApplication
+import com.mohamed.medhat.graduation_project.dagger.components.ActivityComponent
 
 /**
  * A base class for all the activities in the app. This class is meant to reduce the boilerplate code for the activity classes.
  */
 open class BaseActivity<V : BaseView> : AppCompatActivity() {
-    lateinit var appComponent: AppComponent
+    lateinit var activityComponent: ActivityComponent
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        appComponent = (application as DaggerApplication).appComponent
+        activityComponent =
+            (application as DaggerApplication).appComponent.activityComponent().create()
         super.onCreate(savedInstanceState)
     }
