+++ app/src/main/java/com/mohamed/medhat/graduation_project/ui/login_activity/LoginViewModel.kt
@@ -1,9 +1,11 @@
 package com.mohamed.medhat.graduation_project.ui.login_activity
 
 import androidx.lifecycle.ViewModel
+import com.mohamed.medhat.graduation_project.dagger.scopes.ActivityScope
 
 /**
  * A ViewModel for the [LoginActivity].
  */
+@ActivityScope
 class LoginViewModel : ViewModel() {
 }
