+++ app/src/main/java/com/mohamed/medhat/graduation_project/ui/splash_activity/SplashActivity.kt
@@ -2,7 +2,6 @@ package com.mohamed.medhat.graduation_project.ui.splash_activity
 
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import com.mohamed.medhat.graduation_project.R
 import com.mohamed.medhat.graduation_project.ui.base.BaseActivity
 import javax.inject.Inject
@@ -18,7 +17,7 @@ class SplashActivity : BaseActivity<SplashView>(), SplashView {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_splash)
-        appComponent.inject(this)
+        activityComponent.inject(this)
         initPresenter(splashPresenter)
     }
