+++ app/src/main/java/com/mohamed/medhat/graduation_project/ui/splash_activity/SplashPresenter.kt
@@ -1,11 +1,13 @@
 package com.mohamed.medhat.graduation_project.ui.splash_activity
 
+import com.mohamed.medhat.graduation_project.dagger.scopes.ActivityScope
 import com.mohamed.medhat.graduation_project.ui.base.BasePresenter
 import javax.inject.Inject
 
 /**
  * An mvp presenter for the [SplashActivity].
  */
+@ActivityScope
 class SplashPresenter @Inject constructor() : BasePresenter<SplashView>() {
 
     private lateinit var splashView: SplashView
